change: simplify sequence loading

+++ code/include/bio/sequence/sequences.h
@@ -10,7 +10,7 @@ extern s64 *g_restrict g_offsets;
 extern char *g_restrict g_letters;
 extern sequence_t *g_restrict g_seqs;
 extern s64 g_alignments;
-extern s32 g_seq_len_max;
+extern size_t g_seq_len_max;
 extern s32 g_seq_n;
 
 bool sequences_load_from_file(void);
